Update patch set 2
Patch Set 2: Presubmit-Ready+1

This started as a small clean-up before I was going to make other changes, but looks like we're returning a lot of things that don't make sense, so I'm attempting to fix it here.

I don't think anything here should be too controversial.  Let me know if my understandings of any of this isn't quite correct, but it is how I understand the usage of users/groups to frameworks to be.

Patch-set: 2
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=+1