Update patch set 1
Patch Set 1:

I've dealt with ltp failures in a way that does not require this change.
https://android-review.googlesource.com/c/platform/test/vts-testcase/kernel/+/986265

This patch could also be changed to add MCL_ONFAULT instead of returning an error. HWASan is fine with MCL_ONFAULT. Still, this could be an unexpected change in behavior for the syscall, and there are almost no users, so I don't mind dropping this patch.

Patch-set: 1
Reviewer: Gerrit User 1005967 <1005967@85c56323-6fa9-3386-8a01-6480fb634889>