Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Awesome! How hard would it be to sort by vtable-ness first? I want to pack relocations if we can do it simply and cheaply.

It's not hard to sort that, but I assume you want to do that for everything? If so, this is probably not the way to go as this requires linking twice. If we can show that's something beneficial in general (which I failed to do for this one and hence this CL), we can see if that can be added to clang.

Patch-set: 1
Reviewer: Gerrit User 1349653 <1349653@85c56323-6fa9-3386-8a01-6480fb634889>