Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+2
> 
> i'd rather just expose the TLS-related stuff in a new platform/ directory or whatever, but that's a lot of work, and this CL is strictly better than what's going on currently so lgtm...
> 
> (should we call it `bionic_libc_tls_headers` though? saves a comment on the caller side :-) )

I did have a look to see if I could just expose one file but that would require moving it into a separate directory so I didn't bother because as you said it is a lot of work. There are other places that use "bionic/libc/private" in their include_dirs so I created something that could be used for all of them for now and whose name matched the contents. If you feel strongly about the name then let me know and I'll change it.

Patch-set: 1
Reviewer: Gerrit User 1039338 <1039338@85c56323-6fa9-3386-8a01-6480fb634889>