Update patch set 4
Patch Set 4:

Should we try to call a non-inlined pthread_cond_clockwait if the platform has it? e.g. using a weak symbol or dlopen? It could give us flexibility to alter the pthread_cond ABI, and it might be helpful for TSAN eventually, which intercepts and reimplements some pthread_* APIs:

https://github.com/llvm-mirror/compiler-rt/blob/8447f30d2ff479b4469246c61d6e3c26ef52ba34/lib/tsan/rtl/tsan_interceptors.cc#L1180

e.g. A TSAN interceptor wouldn't be able to intercept pthread_cond_clockwait if it's inlined into libc++_shared.so.

We use a weak symbol for __cxa_thread_atexit -> __cxa_thread_atexit_impl in libc++abi. It's harder to do that here because the names are the same, but maybe it's possible using two asm("...") specifiers? https://godbolt.org/z/VM-I-O

Patch-set: 4
Reviewer: Gerrit User 1229339 <1229339@85c56323-6fa9-3386-8a01-6480fb634889>