Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Seems reasonable.
> 
> > Hit rate seen is mostly within 15% to 45%.
> 
> Is that after applying https://reviews.llvm.org/D65242 to lld?

This is before applying that. With that change, this becomes much more beneficial. For example, running audioserver on an ARM32 device, the number of cache hit goes up from 20760 to 63279, out of 110729 symbol lookups. Similarly for cameraserver, there are a total of 89015 symbol lookups, and the hit goes up from 11003 to 46495.

Patch-set: 2
Reviewer: Gerrit User 1349653 <1349653@85c56323-6fa9-3386-8a01-6480fb634889>