Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (that "arm64" code is the 32-bit vdso for a 64-bit kernel :-) )
> 
> On Wed, Jul 31, 2019 at 5:38 PM Alistair Delva (Gerrit)

Right, which is used by any 32-bit app running on said kernel. So a still a big deal (zygote32 no?) My comment was referring to the context that we had only seen this on cuttlefish so far, so thought it might be x86-specific. Alas, no, likely not..

Patch-set: 1
Reviewer: Gerrit User 1293738 <1293738@85c56323-6fa9-3386-8a01-6480fb634889>