Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> i already did. please put all of this into one file.

Hi, we have plans to update the support for expf component as well.Currently we are validating the performance on that one.
Both exp and expf and will refer to the same cout and data files, so we are planning  to keep them separate from their actual implementations.

Patch-set: 1
Reviewer: Gerrit User 1275029 <1275029@85c56323-6fa9-3386-8a01-6480fb634889>