fortify: use __builtin_dynamic_object_size
`__builtin_dynamic_object_size` is a more powerful form of
`__builtin_object_size`. In particular, it's not absolutely required to
fold to a constant, so it allows for us to also detect the size of
allocations produced by constructs like `malloc(n)`, where `n` is
non-constant.

To be clear, constructs like the `malloc` mentioned above still have to
be locally visible; we're not storing meaningful amounts of data on the
side like e.g. ASAN does. This new spelling of `__builtin_object_size`
just allows us to take advantage of a bit more local information, even
if that information doesn't compile down into a constant.

This also adds a few #defines so we can trivially opt out of this if
necessary, similar to how we have _FORTIFY_SOURCE=1, etc.

Bug: 131861088
Test: checkbuild on blueline internal-master; blueline boots.
Change-Id: I0c299bd9442cf512781c7032707f08133f5cb01b