Update patch set 2
Patch Set 2:

interesting --- this seems to break the malloc_usable_size tests, because clang doesn't believe you should be able to take advantage of that.

(unrelated: we can remove the diagnostics but we still need to keep the translation to the __builtin_..._chk functions? clang doesn't know how to do that itself?)

Patch-set: 2
CC: Gerrit User 1003224 <1003224@85c56323-6fa9-3386-8a01-6480fb634889>