Update patch set 3
Patch Set 3:

(2 comments)

Thanks for submitting a patch and taking care of Treble! Let me first try to understand what you are willing to achieve with this. It looks like you are working on to do the system-only upgrade for the devices using VNDK-lite. Am I correct?

If so, I am afraid to say that this wasn't our goal. VNDK-lite, as you might know, was our short-term solution for the devices using old BSP that the SoC vendors don't (or can't) support. In our case, it was the original Pixel. Since VNDK is not fully enforced on these devices, something in vendor partition is very likely to have dependencies on the system private libraries which we don't keep stable across releases. So, by design, system-only upgrade is not "guaranteed" to be successful on such devices. It's good to know that the Xiaomi device was bootable with the change, but it might fail on some operations (like telephony, media play, etc.)

However, I actually am fine with doing something like this to at least get a chance.

I just wanted to make it clear that this path (system-only OTA with vndk-lite) is not guaranteed to be successful and thus is not officially supported.

Patch-set: 3
Reviewer: Gerrit User 1132673 <1132673@85c56323-6fa9-3386-8a01-6480fb634889>