Update patch set 1
Patch Set 1:

(11 comments)

I'm not really a NEON expert either, but the code did look reasonable. One minor note is that it might be good to point to or replicate some documentation about the GNU hash function in the assembly source file itself. I know it always helps me to see the pseudocode for some ASM before I try to interpret it in my head.

Patch-set: 1