Do not provide nonfunctioning placeholders for the libdl functions when linking a static executable.
Replaces silent misbehaviour at runtime with a build time error, which makes
mistakes due to this much easier to detect.

This means that code that need to work in a static executable either has to
use cpp macros to provide an alternative implementation, or link with
libdl_static_stubs to get the placeholders into the executable.

Test: Build & boot
Test: m checkbuild
Test: m bionic-unit-tests-static && \
  adb push out/target/product/taimen/testcases/bionic-unit-tests-static /data && \
  adb shell data/bionic-unit-tests-static/arm/bionic-unit-tests-static && \
  adb shell data/bionic-unit-tests-static/arm64/bionic-unit-tests-static
Test: atest CtsBionicTestCases
Bug: 141485154
Change-Id: Ia37cd89a8dab0955c6c899772bbbf7a04ed7ff83