Do not provide nonfunctional replacements for the libdl functions when linking a static executable.
This way we get a build time error instead of misbehaviour at runtime, which
makes mistakes due to this much easier to detect, in particular for
libraries that may be statically linked, but which require dlopen etc and
should not be allowed in static executables.

This means that code that need to work in a static executable either has to
use cpp macros to provide an alternative implementation, or link with
libdl_static_dummy to get the dummy dl* replacements into the executable.

Test: Build & boot
Test: m checkbuild
Test: atest CtsBionicTestCases
Bug: 141485154
Change-Id: Ia37cd89a8dab0955c6c899772bbbf7a04ed7ff83