Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > I'm a bit uneasy about the lock upgrade API, for a few reasons. I haven't attempted to understand the linker side of this yet.
> > 
> > 1. This is really tricky stuff. I would expect that Posix considered such an API. I know that WG21 did (e.g. wg21.link/N2094). WG21 never ended up comfortable enough with it to take theirs.
> 
> Sure. Someone has to go first.

Yes. But the fact that discussions in larger groups generally seem to have decided not to go ahead suggests caution.

> 
> > 2. This doesn't seem to play well with recursive reader locks, which are required by Posix. 
> 
> Can you point out the cases where we'd break recursive reader locks? On the downgrade path, there are no readers. On the upgrade path, we wait for all readers to finish, meaning that if the current thread owns multiple reader locks, we deadlock. That's a limitation of the API spelled out in the documentation.

I actually didn't see it spelled out in the documentation. That could be fixed. But I think it's worse than that. This is effectively a writer-priority lock acquisition. Consider:

Thread 1:          rdlock; upgrade

Thread 2: rdlock;                   rdlock;

I believe this deadlocks as well. Thus I don't think this is compatible with reentrant lock acquisition at all.

I think the only reasonable way to handle this is to restrict the new functionality to PTHREAD_RWLOCK_PREFER_WRITER_NONRECURSIVE_NP rwlocks. That also avoids an implementation that's half reader-, half writer-preference.

> 
> > (Admittedly Posix also seems to have missed the issue with reader recursion and writer preference. That doesn't add to my confidence in this area.)
> 
> POSIX doesn't get everything right the first time. :-)
> 
> > 3. I think this solution is only useful if some writers do so by upgrading, and others write directly. If all writes did so by upgrading, and you knew that there is only one writer at a time, you could just drop the reader lock and reaquire it as a writer. 
> 
> The problem is that if you drop the reader lock and re-acquire it as a writer, you can observe arbitrary state changes between the reader-drop and the writer-acquire. The upgrade mechanism excludes these changes.

Yes, but you're already assuming that there's only one thread that can possibly upgrade at any given time. If you also assume that all writers do so by upgrading, then no writer can sneak in between releasing the read lock and acquiring the write lock.

> 
> > If you have a mixture of writer types, you could presumably still have a separate "I might write" mutex that's acquired by writers, and by potentially upgrading readers, before acquiring the rwlock. That probably costs you some concurrency. This makes it seem like a bit of a niche solution.
> 
> I considered a two-lock approach. The problem is that upgrading is rare, and I don't want to nearly double the locking overhead on every entry into the linker. There are also logic issues: keep in mind that we need to be able to nest dlopen arbitrarily deep and that each dlopen needs to go through a phase that excludes concurrent readers and another phase that allows concurrent readers. The upgrade-downgrade approach models exactly the mutual exclusion we want and it does so while imposing no overhead (except one well-predicted branch) on the common case.

It's important that we be able to upgrade in the presence of concurrent writers, and that we must always be able to do so?

> 
> > If there's another way to solve the linker problem, I would probably be more comfortable with that. Conversely, an argument that that problem can only be solved in this manner would strengthen the argument for adding the API.
> 
> The other approach is a linker refactor that adds fine-grained per-DSO locking. I think that's too big of a project right now. This CL gets us concurrency that helps in real-world scenarios (e.g., camera startup) without having to rearchitect the linker or accept a performance hit on paths that are already very hot.

The refactoring does indeed sound scary, particularly for something like dl_iterate_phdr().

Is there a way to keep the functionality private to the linker, at least for now? I don't object to the approach there. My concern is about introducing a public API in a reasonably-well-explored tricky area, without a lot of usage experience. If we get this wrong, we have to maintain a broken API

Patch-set: 3