Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> do you know _why_ you regressed so badly with this kernel? is this spectre mitigations or something?

Devices were with 3.18 kernel and now they upreved to 4.19 so hard to say what exactly caused it. This test can successfully pass with 70 threads so we can use 70 or do 50 to avoid future issues. I don't think there are a lot of devices with 2 cores and Android P.

Patch-set: 2