[bionic][openbsd] fix -Wimplicit-fallthrough
Clang r367134 (clang-10) added support for the GNU C style attribute for
denoting intentional fallthrough between switch statements, and started
warning on -Wimplicit-fallthrough for C code. Previously, this warning
was disabled for C as there was no way to denote to Clang that the
fallthrough was intentional.

Use:
__attribute__((__fallthrough__));
before a case statement to denote intentional fallthrough. Support for
the attribute can be feature detected via the use of __has_attribute.

Bug: 139945549
Test: mm
Change-Id: Ia9a40270ed24bd28739b3860270633230a771e2b
Signed-off-by: Nick Desaulniers <ndesaulniers@google.com>