[bionic] disable warnings for clang-370808 upgrade
Sets:
-Wno-implicit-fallthrough
-Wno-reorder-init-list
-Wno-implicit-int-float-conversion

Clang r367134 (clang-10) added support for the GNU C style attribute for
denoting intentional fallthrough between switch statements, and started
warning on -Wimplicit-fallthrough for C code. Previously, this warning
was disabled for C as there was no way to denote to Clang that the
fallthrough was intentional.

The upstream {free|net|open}bsd code needs to be fixed.

struct sigaction's member ordering differs in its definition based on
__LP64__, which makes writing portable designated initializers ugly.

The imprecision within libm should be resolved upstream with a proper
solution backported after the compiler upgrade.

Bug: 139945549
Bug: 143311272
Test: mm
Change-Id: I2f2512895e4e5cac74e5ff04c04e6b71444d2f9f
Signed-off-by: Nick Desaulniers <ndesaulniers@google.com>