Update patch set 3
Patch Set 3:

> Patch Set 3: Presubmit-Ready+1
> 
> i still think something's badly broken here, and you should probably look into finding the real problem. but the patch i just uploaded is slightly cleaner than what's currently checked in, and i doubt anyone else is doing anything with 2-core devices, so i don't really object, but...
> 
> +cferris to have a look to see if we're accidentally running too many tests at once before we submit this. (because ideally we should just be able to remove this hack entirely. it seems unlikely to me that you can't actually run the 100-thread variant of this test on a 2-core device. i'm assuming that if you run this test directly with `adb shell /data/nativetest/bionic-unit-tests/bionic-unit-tests --gtest_filter='*leak*'` it passes even without the hack, right?)

I verified that the default number of concurrent tests default to sysconf(_SC_NPROCESSORS_ONLN). Is it possible that your new kernel is returning data that fools the code into thinking there are more processors than actually exists? Or even the older version did the same?

The code reads the output of /sys/devices/system/cpu and counts the number of lines that start with cpu[0-9]+.

We have had similar problems in the past resulting from running too many tests concurrently, so it would be useful to see if too many tests are running at the same time.

Patch-set: 3