Fix allocations escaping malloc debug.
When using a FILE object for some malloc debug functions which trigger
allocations being made in the object when calling fprintf. The problem
is that these allocations are not malloc debug allocations so they
survive in the FILE object can get freed with malloc debug enabled
again. In many cases, the code will detect the bad pointer and leak
the memory, but it might also cause a crash.

The fix is to avoid using fprintf so that no allocations are made
in the object that survive and need to be freed in the fclose call.

Change the MallocXmlElem.h to use a file decsriptor not a FILE*.

Add new unit and system tests to detect this case.

Bug: 143742907

Test: Ran unit and system tests.
Change-Id: I524392de822a29483aa5be8f14c680e70033eba2