WIP: Add a hook to track file descriptor life cycles.
Add a hook that's called upon file descriptor creation to enable the
fetching of backtraces to track down file descriptor leaks.

Currently, this is not even close to being complete, but completeness
isn't required for this to be useful, and most of the other sources of
file descriptors are implementable with trivial wrappers. The main
exception is binder, which receives file descriptors as a payload in a
not-trivially-parsable byte blob. Adding parsing of the (unstable)
binder wire protocol to libc seems undesirable, so the best approach
might be to expose the hook for other libraries to call.

TODO: dup(2,3), fcntl F_DUPFD(_CLOEXEC), epoll, inotify
TODO: tests, documentation
TODO: how do we turn this on?

Bug: http://b/140703823
Test: TODO
Change-Id: I308a14c2e234cdba4207157b634ab6b8bc539dd9