Add automatic running of tests on linker changes.
malloc debug and malloc hooks have been broken for a long time
and no one noticed. So add them to be run by default on linker
changes since that is where a problem is likely to occur.

Also, changed the names of the malloc hooks tests to system tests
because they weren't really unit tests.

Test: Ran tests.
Change-Id: Idea4c864f1d62598148ee78d7c9397e45234b1ca