WIP: Optimize GNU hash linking for large inputs
Symbol lookup is O(L) where L is the number of libraries to search (e.g.
in the global and local lookup groups). Factor out the per-DSO work into
two functions, soinfo_do_lookup_impl and finish_lookup, and optimize for
the situation where all the DSOs are using DT_GNU_HASH (rather than SysV
hashes).

To load a set of libraries, the loader first constructs an auxiliary list
of libraries (LookupList, containing LookupLib objects). The LookupList is
reused for each DSO in a load group. (-Bsymbolic is accommodated by
modifying the LookupLib at the front of the list.) To search for a symbol,
soinfo_do_lookup_impl iterates over the list of libraries scanning GNU
bloom filters, and then it calls the (inlined) finish_lookup to scan the
GNU hash table.

There was a slight improvement from templatizing soinfo_do_lookup_impl
and skipping the does-this-DSO-lack-GNU-hash check. (I had been wondering
if it would help to prefetch entries from the bloom filters, but I didn't
see an improvement the first time I tried that.)

The initial intent of having an auxiliary list was to avoid having to
check the soinfo version field, but maybe we could just use a vector of
soinfo* and rely on FLAG_GNU_HASH to enter the SysV-lookup code path.

Add an assembly version of the GNU symbol hash calculation that uses
vector instructions.

I've measured this CL reducing the initial link time of large programs by
25-40% (e.g. audioserver, cameraserver, etc).

Test: bionic-unit-tests
Change-Id: If40a42fb6ff566570f7280b71d58f7fa290b9343