Integrate GWP-ASan into bionic's malloc().
This patch introduces GWP-ASan - a sampled allocator framework that
finds use-after-free and heap-buffer-overflow bugs in production
environments.

GWP-ASan is being introduced in an always-disabled mode. This means that
GWP-ASan will be permanently disabled until a further patch turns on
support. As such, there should be no visible functional change for the
time being.

As part of this patchset, we also move malloc_common.cpp into its own
static library that's linked everywhere it should. This allows us to
build just the allocation functions with -fno-emulated-tls. Platform
(non-emulated TLS) is a requirement of GWP-ASan.

More information about GWP-ASan can be found in the upstream
documentation: http://llvm.org/docs/GwpAsan.html

Bug: 135634846
Test: atest bionic

Change-Id: If32a3d24cb54f08091f96f06e8e35debc30a253b