Create a libunwind.so shared library containing the unwinder implementation.
This supports the soong commit which causes most platform binaries to
dynamically link against libunwind.so instead of statically linking against
the unwinder implementation. The soong commit message has more motivation
for this change.

ARM32 uses LLVM libunwind, while all other platforms use libgcc as the
unwinder implementation. This matches the current behaviour.

On ARM32, the library libc.so already exports an unwinder implementation
(previously libgcc), which we need to keep exporting for app compatibility
reasons. In order to avoid ambiguity between the libc.so unwinder and the
libunwind.so unwinder, the libc.so unwinder symbols are made to forward to
libunwind.so, which means that apps which were directly linking against the
libc.so unwinder symbols are now using LLVM libunwind instead of libgcc.

Bug: 144430859
Change-Id: I52c7f7893d93f500383aeb0b76086c3b6f1935a5